+++ bucketsort.h
@@ -1,6 +1,6 @@
 #ifndef BUCKETSORT_H
 #define BUCKETSORT_H
 
-void bucketSort(int* vet, int tam, int numeroBaldes, int* comp, int* troca);
+void bucketSort(int* vet, int tam, int numeroBaldes, long* comp, long* troca);
 
 #endif
