+++ insdireta.h
@@ -1,6 +1,6 @@
 #ifndef INSDIRETA_H
 #define INSDIRETA_H
 
-void insercaoDireta(int* vet, int tam, int* troca, int* comp);
+void insercaoDireta(int* vet, int tam, long* troca, long* comp);
 
 #endif
