+++ quicksortfim.h
@@ -1,6 +1,6 @@
 #ifndef QUICKSORTFIM_H
 #define QUICKSORTFIM_H
 
-void quickSortFim (int* vet, int esq, int dir, int* comp, int* troca);
+void quickSortFim (int* vet, int esq, int dir, long* comp, long* troca);
 
 #endif
