+++ shellsort.h
@@ -1,6 +1,6 @@
 #ifndef SHELLSORT_H
 #define SHELLSORT_H
 
-int* shellSort (int* vet, int tam, int* troca, int* comp);
+int* shellSort (int* vet, int tam, long* troca, long* comp);
 
 #endif
